Use custom gesture recognizer instead of UIPanGestureRecognizer
This change fixes several bugs:

* Pen tool would ignore the first ~30px of your touch
* Pen shapes would have lower line segment resolution for the first ~30 px even after starting
* Text handles would sometimes fail to recognize drags that started on the handles

The solution is to  replace UIPanGestureRecognizer with our own custom internal implementation that simply tracks raw touch points without trying to do anything fancy. When the touch has ended, DrawsanaView checks whether the touch moved more than 10px, and if it did, it cancels the pan on the tool and instead calls it a tap.